requirements.txt
celery==5.2.3
Django==3.2.9
djangorestframework==3.13.0
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
factory-boy==3.2.0
flower==1.0.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pytest-django==4.1.0
redis==4.1.4
